// frontend/src/app/models/user.model.ts
export interface User {
  id?: number;
  email: string;
  username?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  username: string;
  password: string;
}

export interface AuthResponse {
  access_token: string;
  user?: User;
}